merge-fix: loop bounds stay 32-bit counts where a program id is now 64-bit (twelve sites, the Metal lowering refuses a 64-bit scf.for bound); a capability test refuses on every backend, not only on Metal
Two semantic conflicts no git conflict signalled, between the trunk's
64-bit program ids (aa60c5c, register 58) and the Mac's 22b427c:

- 22b427c says the Metal induction lowering does not terminate on a 64-bit
  `scf.for` bound and narrows repeat_interleave's count. With every program
  id 64-bit, twelve loop bounds derived from one (cumsum, decode_attn_vec,
  flash_decode, sort_op's tile ranges, the two upsample_nearest counters)
  would reach Metal as 64-bit. Each bound scalar is narrowed with
  .to(tl.int32) where it is defined; addressing keeps the 64-bit program
  ids. tests/unit/kernels/test_loop_bounds_are_32_bit.py reads the source
  and refuses a bound derived from a program id without the narrowing
  (seen RED on cumsum's block_offset with the cast removed). CUDA-measured
  here (the kernel suite and the byte pair follow); the Metal side is the
  Mac's battery to confirm.
- test_a_backend_without_a_row_refuses built a table WITH a cuda row and
  expected a refusal: green on Metal, red on CUDA — a test at a value that
  makes the branch equivalent on one backend. The table now names no
  existing backend, so it refuses everywhere.

+++ src/neurobrix/kernels/ops/cumsum.py
@@ -186,8 +186,8 @@ def reduce_then_scan_block_sum_kernel_row(
     pid_n = tl.program_id(1).to(tl.int64)
     pid_m = tl.program_id(0).to(tl.int64)
     num_programs_n = tl.num_programs(1)
-    block_offset = pid_n * (tiles_per_cta * TILE_SIZE)
-    block_end = min(block_offset + tiles_per_cta * TILE_SIZE, N)
+    block_offset = (pid_n * (tiles_per_cta * TILE_SIZE)).to(tl.int32)   # loop bounds are 32-bit counts: the Metal lowering refuses a 64-bit scf.for bound (addressing stays 64-bit through the program ids)
+    block_end = min(block_offset + tiles_per_cta * TILE_SIZE, N).to(tl.int32)
 
     acc_dtype: tl.constexpr = get_scan_accum_type(in_ptr.type.element_ty)
     acc = tl.zeros((TILE_SIZE,), dtype=acc_dtype)
@@ -226,8 +226,8 @@ def reduce_then_scan_block_scan_kernel_row(
     """Final block scan: apply prefix from previous blocks, then local cumsum."""
     pid_m = tl.program_id(0).to(tl.int64)
     pid_n = tl.program_id(1).to(tl.int64)
-    block_offset = pid_n * (tiles_per_cta * TILE_SIZE)
-    block_end = min(block_offset + tiles_per_cta * TILE_SIZE, N)
+    block_offset = (pid_n * (tiles_per_cta * TILE_SIZE)).to(tl.int32)   # loop bounds are 32-bit counts: the Metal lowering refuses a 64-bit scf.for bound (addressing stays 64-bit through the program ids)
+    block_end = min(block_offset + tiles_per_cta * TILE_SIZE, N).to(tl.int32)
     acc_dtype: tl.constexpr = get_scan_accum_type(in_ptr.type.element_ty)
 
     prefix = tl.load(
